% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{basal_metabolic_rate}
\alias{basal_metabolic_rate}
\title{Basal metabolic rate}
\usage{
basal_metabolic_rate(body_weight, body_height, age, sex = c("male", "female"))
}
\arguments{
\item{body_weight}{Body weight in kg}

\item{body_height}{Body height in m}

\item{age}{Athlete's age in years}

\item{sex}{Select sex of athlete (male or female).}
}
\description{
Estimate the basal metabolic rate using the Harris-Benedict equation.
}
