% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{body_fat}
\alias{body_fat}
\title{Body fat}
\usage{
body_fat(
  body_height,
  abdominal_circumference,
  neck_circumference,
  sex = c("male", "female")
)
}
\arguments{
\item{body_height}{Body height in centimeters}

\item{abdominal_circumference}{Abdominal circumference in centimeters.}

\item{neck_circumference}{Neck circumference in centimeters.}

\item{sex}{Select sex of athlete (male or female).}
}
\value{
Returns a vector containing estimated bodyfat and 95 \% confidence intervals.
}
\description{
Estimate body fat percentage using the circumference-based approach
by Lahav et al., 2018 published in the British Journal of Nutrition.
This function estimates the body fat percentage and 95\% confidence intervals.
}
\details{
Lahav Y, Epstein Y, Kedem R, Schermann H. A novel body circumferences-based estimation of percentage body fat. Br J Nutr. 2018 Mar;119(6):720-725. doi: 10.1017/S0007114518000223. PMID: 29553036.
}
