% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{.md_to_html}
\alias{.md_to_html}
\title{Internal: convert Markdown to gridtext-compatible HTML}
\usage{
.md_to_html(md, link_color = "#2563EB")
}
\description{
Internal: convert Markdown to gridtext-compatible HTML
}
\keyword{internal}
