% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{.subtitle_cell}
\alias{.subtitle_cell}
\title{Internal: subtitle cell with gradient and rounded corners}
\usage{
.subtitle_cell(
  label,
  cell_bg_cols = c("#2f6cab", "#173052"),
  cell_bg_stops = NULL,
  cell_bg_dir = "lr",
  box_r = unit(10, "pt"),
  box_border_col = NA,
  box_border_lwd = 1,
  text_gp = gpar(col = "white", fontsize = 16, fontfamily = "Noto Sans", fontface =
    "bold"),
  text_pad = unit(c(10, 14, 10, 14), "pt"),
  box_margin = unit(c(6, 6, 6, 6), "pt")
)
}
\description{
Internal: subtitle cell with gradient and rounded corners
}
\keyword{internal}
