% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{ideal_weight}
\alias{ideal_weight}
\title{Ideal weight}
\usage{
ideal_weight(lean_body_mass, desired_body_fat)
}
\arguments{
\item{lean_body_mass}{Lean body mass in kg}

\item{desired_body_fat}{Desired body fat percentage e.g., 12 \%.}
}
\description{
Estimate ideal weight at preferred body fat percentage
}
