% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{lean_body_mass}
\alias{lean_body_mass}
\title{Lean body mass}
\usage{
lean_body_mass(body_weight, body_fat)
}
\arguments{
\item{body_weight}{Body weight in kg}

\item{body_fat}{Body fat in percent}
}
\description{
Lean body mass
}
