% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{new_canvas}
\alias{new_canvas}
\title{Create a new canvas gtable}
\usage{
new_canvas()
}
\value{
A `gtable` object representing an empty canvas.
}
\description{
Initializes an empty vertical stack to which rows can be appended.
}
