% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{norm_volume_index}
\alias{norm_volume_index}
\title{Normalized volume index}
\usage{
norm_volume_index(repetitions, total_load, body_weight)
}
\arguments{
\item{repetitions}{Total repetitions of a particular exercise.
This is classically the product of the number of reps and the number
of working sets per training session e.g., 3 sets of 10 reps = 30 total repetitions}

\item{total_load}{Total amount of weight moved e.g., a squats session
of 3 sets with 10 reps with 120 kg the total_load = 3 x 10 x 120.}

\item{body_weight}{Body weight in kg}
}
\description{
Calculate the normalized volume index for resistance exercises
}
