% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{onerm_brzycki}
\alias{onerm_brzycki}
\title{1RM after Brzycki}
\usage{
onerm_brzycki(load, repitions)
}
\arguments{
\item{load}{Weight used during the exercise in kg.}

\item{repetitions}{Number of repetitions.}
}
\description{
Calculate the 1RM
}
