% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{onerm_epley}
\alias{onerm_epley}
\title{1RM after Epley}
\usage{
onerm_epley(load, repitions)
}
\arguments{
\item{load}{Weight used during the exercise in kg.}

\item{repetitions}{Number of repetitions.}
}
\description{
Calculate the 1RM
}
