% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{onerm_oconner}
\alias{onerm_oconner}
\title{1RM after O'Conner}
\usage{
onerm_oconner(load, repetitions)
}
\arguments{
\item{load}{Weight used during the exercise in kg.}

\item{repetitions}{Number of repetitions.}
}
\description{
Calculate the 1RM
}
