% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{onerm_wathen}
\alias{onerm_wathen}
\title{1RM after Wathen}
\usage{
onerm_wathen(load, repetitions)
}
\arguments{
\item{load}{Weight used during the exercise in kg.}

\item{repetitions}{Number of repetitions.}
}
\description{
Calculate the 1RM
}
