% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{to_in}
\alias{to_in}
\title{Convert `grid::unit` to inches}
\usage{
to_in(u)
}
\arguments{
\item{u}{A `grid::unit` object.}
}
\value{
Numeric length in inches.
}
\description{
Robust conversion helper for unit objects.
}
