% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{training_intensity}
\alias{training_intensity}
\title{Training Intensity}
\usage{
training_intensity(sets, repetitions, load)
}
\arguments{
\item{sets}{Number of sets of an exercise per training session.}

\item{repetitions}{Number of repetitions per set.}

\item{load}{Weight used during the exercise in kg.}
}
\description{
Calculate the training intensity.
}
